Fix review findings: Permute3D, PixtralImageSizes, config_filename, DRY
Critical:
- Add missing Permute3D + PixtralImageSizes transforms to Pixtral
  pipeline (5 steps → 7 steps, matching olive-recipes reference)
- Fix config_filename mismatch: Pixtral with_vision() now passes
  config_filename="processor_config.json" (was using default
  "image_processor.json")

Medium:
- Extract _build_vision_transform_pipeline() shared helper to DRY
  the Pixtral and generic VLM branches (~80 duplicated lines removed)

Low:
- Add mistral3 to _ORT_GENAI_MODEL_TYPE + _PIXTRAL_MODEL_TYPES
- Update module docstring to mention processor_config.json
- Improve pixtral test: verify 7-step pipeline, all transform types
  in order, Permute3D dims, and output filename

+++ examples/olive/ministral-3-3b-vlm/eval.py
@@ -183,7 +183,7 @@ def build_pytorch_runner(model_id: str):
     print(f"  Device: {device}, dtype: {dtype}")
 
     pt_model = Mistral3ForConditionalGeneration.from_pretrained(
-        model_id, dtype=dtype, trust_remote_code=True
+        model_id, torch_dtype=dtype, trust_remote_code=True
     ).to(device)
     pt_proc = AutoProcessor.from_pretrained(model_id, trust_remote_code=True)
     print("  PyTorch model loaded.")
